first running example for wall-et lift adjustment process

+++ controlcomponents/de.dfki.cos.basys.p4p.controlcomponent.wallet/src/main/java/de/dfki/cos/basys/p4p/controlcomponent/wallet/LiftOperationMode.java
@@ -23,7 +23,7 @@ import de.dfki.cos.basys.controlcomponent.annotation.OperationMode;
 public class LiftOperationMode extends BaseOperationMode<WalletService> {
 	
 	@Parameter(name = "level", direction = ParameterDirection.IN)
-	private int level;
+	private long level;
 	private long startTime = 0;
 	private long endTime = 0;
