+++ controlcomponents/de.dfki.cos.basys.p4p.controlcomponent.wallet/src/main/java/de/dfki/cos/basys/p4p/controlcomponent/wallet/WalletService.java
@@ -8,7 +8,7 @@ public interface WalletService {
 	
 	void gotoSymbolicPosition(String positionName);
 	
-	void moveLiftToLevel(int level);
+	void moveLiftToLevel(long level);
 	
 	GoalStatusEnum getStatus();
